Add LevelDBStoredSortedMap an implementaion of SortedMap on LevelDB
This implementation currently passes the SortedMapInterfaceTest in
Guava. It still misses documentation and needs to be tested in more
cases then what the Guava test does.

Using this class is currently also a bit complicated since it
needs the user to provide both the DB, DBComparator and optionally
a normal Comparator. The user must make sure that the DBComparator
is set correctly on the DB and that it matches what either the
Comparator or the Comparable does.